+++ lib/vector_embedding/index.js
