+++ lib/vector_embedding/semantic-search/embedding.js
