+++ lib/vector_embedding/semantic-search/model-utils.js
