+++ package.json
@@ -8,10 +8,6 @@
   "author": "SAP SE (https://www.sap.com)",
   "homepage": "https://cap.cloud.sap/",
   "main": "cds-plugin.js",
-  "exports": {
-    ".": "./cds-plugin.js",
-    "./vector-embedding": "./lib/vector_handling/sync-wrapper.js"
-  },
   "scripts": {
     "lint": "npx -y eslint@10 .",
     "test": "node --test tests/*.test.js",
